Büyük güncelleme.

+++ DataAccess/Concrete/EntityFramework/EfProductDal.cs
@@ -21,7 +21,7 @@ namespace DataAccess.Concrete.EntityFramework
 				var result = from p in context.Products
 							 join c in context.Categories
 							 on p.CategoryId equals c.CategoryId
-							 select new ProductDetailDto { ProductId = p.ProductId, ProductName = p.ProductName, CategoryName = c.CategoryName, UnitsInStock = p.UnitsInStock };
+							 select new ProductDetailDto { ProductId = p.ProductId, ProductName = p.ProductName, CategoryName = c.CategoryName, UnitsInStock = (short)p.UnitsInStock };
 
 				return result.ToList();
