+++ Entities/Concrete/Product.cs
@@ -12,8 +12,9 @@ namespace Entities.Concrete
         public int ProductId { get; set; }
         public int CategoryId { get; set; } 
         public string  ProductName { get; set; }
-        public short UnitsInStock { get; set; }
+        public int UnitsInStock { get; set; }
         public decimal UnitPrice { get; set; }
+
 
     }
 }
